+++ packages/core/src/Icon.js
@@ -1,28 +1,33 @@
 import React from 'react'
-import styled from 'styled-components'
-import { space, size, color } from 'styled-system'
-import Tag from 'clean-tag'
 import PropTypes from 'prop-types'
+import styled from 'styled-components'
+import { space, color } from 'styled-system'
 import icons from '../icons.json'
 import theme from './theme'
 
 const getPath = ({ name }) => icons[name]
 
-const Base = ({ name, size, ...props }) => {
+const Svg = styled.svg`
+  flex: none;
+  ${space} ${color};
+`
+
+const Icon = ({ name, size, ...props }) => {
   const icon = getPath({ name })
   if (!icon) return false
 
   return (
-    <Tag is="svg" {...props} viewBox={icon.viewBox} fill="currentcolor">
+    <Svg
+      {...props}
+      viewBox={icon.viewBox}
+      width={size}
+      height={size}
+      fill="currentcolor"
+    >
       <path d={icon.path} />
-    </Tag>
+    </Svg>
   )
 }
-
-const Icon = styled(Base)`
-  flex: none;
-  ${space} ${size} ${color};
-`
 
 Icon.displayName = 'Icon'
 
@@ -37,16 +42,16 @@ const keys = Object.keys({
 })
 
 Icon.propTypes = {
+  size: PropTypes.number,
   ...space.propTypes,
+  ...color.propTypes,
   name: ({ name }) => {
     if (!keys.includes(name)) {
       return new Error(
         `Failed prop type: Invalid prop name of value '${name}' supplied to Icon, expected one of ${keys.toString()} is expected`
       )
     }
-  },
-  size: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-  color: PropTypes.string
+  }
 }
 
 export default Icon