+++ packages/core/src/Select.js
@@ -36,8 +36,13 @@ SelectBase.defaultProps = {
   py: 14
 }
 
+SelectBase.propTypes = {
+  ...space.propTypes,
+  ...fontSize.propTypes
+}
+
 const Select = styled(props => (
-  <Flex width={1} align="center">
+  <Flex width={1} alignItems="center">
     <SelectBase {...props} />
     <ClickableIcon ml={-32} name="chevronDown" color="gray" />
   </Flex>