+++ packages/core/storybook/Colors.js
@@ -2,9 +2,9 @@ import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { Box, Flex, Text, theme } from '../src'
 
-const keys = Object.keys(theme.colors).filter(
-  key => !Array.isArray(theme.colors[key])
-)
+const blacklist = ['lightOrange', 'darkPurple']
+
+const keys = Object.keys(theme.colors).filter(key => !blacklist.includes(key))
 
 const next = keys.map(key => ({ key, value: theme.colors[key] }))
 
@@ -26,14 +26,11 @@ storiesOf('Color', module).add('Palette', () => (
       <h1>Color Palette</h1>
     </Box>
     <Flex wrap>
-      {next.map(
-        color =>
-          !color.key[color.key.length - 1].match(/^\d+$/) ? (
-            <Box key={color.key} p={3} width={[1, 1 / 2, 1 / 3, 1 / 4, 1 / 5]}>
-              <Card name={color.key} color={color.value} />
-            </Box>
-          ) : null
-      )}
+      {next.map(color => (
+        <Box key={color.key} p={3} width={[1, 1 / 2, 1 / 3, 1 / 4, 1 / 5]}>
+          <Card name={color.key} color={color.value} />
+        </Box>
+      ))}
     </Flex>
   </div>
 ))