+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Update to styled-system v3 #356
+- Adjust styles for `IconButton` #362
+- Update colors #355
+
 ## [2.0.0-0] 2018-09-14
 
 ### Added